requirements.txt
python-telegram-bot==21.3
pyexcel==0.7.0
pyexcel-xlsx==0.6.0
openpyxl==3.0.10
python-dotenv==1.0.1
black==24.4.2
